requirements.txt
# SuperMechs framework
git+https://github.com/Enegg/python-socketio
attrs == 22.1.*
aiohttp == 3.8.*
Pillow == 9.1.*

# Bot
git+https://github.com/Enegg/disnake
certifi == 2021.10.8

# General
python-dotenv == 0.19.*
typing-extensions == 4.2.0